Use adaptor logic from musig2 crate instead of own algo
Use the more obviously correct method to create a link betwwen the swap
tx signature and the seller's private key share for the buyer payout,
which is to jointly form an adaptor signature for the swap tx in place
of an ordinary multiparty signature, using the corresponding pubkey
share directly as the _adaptor point_. This is superior to the naive
method of taking the difference between a partial sig & private key
share, since it doesn't involve the private key at all (except for
signing) and thus provably won't leak any information about it to the
buyer, until the swap tx is published. Conversely, the buyer can be
certain that having a _valid_ adaptor signature (shared with the
seller) means that there is no (ordinary, valid) swap tx signature the
seller can create besides the one that reveals the adaptor secret (that
is, the missing private key share on the buyer payout).

This simplifies the 'PartialSignaturesMessage' proto definition a
little, since the traders are now just exchanging 3 or 4 plain partial
signatures (scalars), to form three ordinary signatures and one (buyer-
side-only) adaptor signature, which is formed in almost exactly the
same way as the other three sigs. (That is, no special 'adaptor' scalar
is being transferred to the buyer.)